Added more comments

+++ BankManager_Csharp/BankManager_Csharp/Models/Transaction.cs
@@ -9,12 +9,22 @@ namespace BankManager_Csharp.Models
     {
         public int transactionId { get; }
         public int accountId { get; }
-        public int balance { get; }
+        public float balance { get; }
         public int amount { get; }
         public String status { get; }
         public DateTime date { get; }
 
-        public Transaction(int transactionId, int accountId, int balance,
+
+        /// <summary>
+        /// Instantiates Transaction object. This is used to record successful transactions.
+        /// </summary>
+        /// <param name="transactionId"></param>
+        /// <param name="accountId"></param>
+        /// <param name="balance"></param>
+        /// <param name="amount"></param>
+        /// <param name="status"></param>
+        /// <param name="date"></param>
+        public Transaction(int transactionId, int accountId, float balance,
             int amount, String status, DateTime date)
         {
             this.transactionId = transactionId;
@@ -26,11 +36,21 @@ namespace BankManager_Csharp.Models
 
         }
 
+
+        /// <summary>
+        /// Returns human-readable datetime
+        /// </summary>
+        /// <returns></returns>
         public String getReadableDate()
         {
             return date.ToString("dd/MM/yyyy hh:mm:ss");
         }
 
+
+        /// <summary>
+        /// Pretty prints transaction information
+        /// </summary>
+        /// <returns></returns>
         public String prettyPrint()
         {
             return "Transaction id: " + transactionId + "\n"
